解决冲突



See merge request !195

+++ index-server/src/main/java/com/cmri/um/he/index/terminal/dao/TerminalOverviewDao.java
@@ -49,7 +49,7 @@ public class TerminalOverviewDao extends BaseDao {
      * @param month
      * @return
      */
-    public int quaryTotal(Integer id, String month) {
+    public double quaryTotal(Integer id, String month) {
         return terminalOverviewMapper.quaryTotal(id,month);
 
     }
